+++ CHANGELOG.md
@@ -8,9 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - [BoxSquaredEL](https://arxiv.org/abs/2301.11118) module added to `mowl.nn`
+- Implemented `model.add_axioms` method. Related to issue #43
 
 ### Changed
 - Bug fixed in corpus generation methods. Issue #36.
+- Updated dependencies to work with Python 3.8 and 3.9. Issue #42
 
 ### Deprecated